test(prepush): move prepush tests to own file
